test: cross-chain complete flow test added, test suite organized
